+++ acme/resources/Factory.py
@@ -7,8 +7,11 @@
 #	Create Resources
 #
 
+"""	This Module provides a resource factory. 
+"""
+
 from typing import Dict, Callable, Tuple, cast
-from ..etc.Types import ResourceTypes as T, addResourceFactoryCallback, FactoryCallableT
+from ..etc.Types import ResourceTypes, addResourceFactoryCallback, FactoryCallableT
 from ..etc.Types import ResponseStatusCode as RC
 from ..etc.Types import Result, JSON
 from ..etc import Utils as Utils
@@ -79,80 +82,85 @@ from ..resources.RBO import RBO
 from ..resources.RBOAnnc import RBOAnnc
 from ..resources.SWR import SWR
 from ..resources.SWRAnnc import SWRAnnc
+from ..resources.WIFIC import WIFIC
+from ..resources.WIFICAnnc import WIFICAnnc
+
 from ..resources.Resource import Resource
 
 
 # Adding factory callbacks to regular resource type details
-addResourceFactoryCallback(T.ACP, 			ACP,			lambda dct, tpe, pi, create : ACP(dct, pi = pi, create = create))
-addResourceFactoryCallback(T.ACPAnnc,		ACPAnnc,		lambda dct, tpe, pi, create : ACPAnnc(dct, pi = pi, create = create))
-addResourceFactoryCallback(T.ACTR, 			ACTR,			lambda dct, tpe, pi, create : ACTR(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.ACP, 			ACP,			lambda dct, tpe, pi, create : ACP(dct, pi = pi, create = create))
+addResourceFactoryCallback(ResourceTypes.ACPAnnc,		ACPAnnc,		lambda dct, tpe, pi, create : ACPAnnc(dct, pi = pi, create = create))
+addResourceFactoryCallback(ResourceTypes.ACTR, 			ACTR,			lambda dct, tpe, pi, create : ACTR(dct, pi = pi, create = create)) 
 # TODO ACTRAnnc
-addResourceFactoryCallback(T.AE, 			AE,				lambda dct, tpe, pi, create : AE(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.AEAnnc,		AEAnnc,			lambda dct, tpe, pi, create : AEAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CIN, 			CIN,			lambda dct, tpe, pi, create : CIN(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CINAnnc, 		CINAnnc,		lambda dct, tpe, pi, create : CINAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CNT,	 		CNT,			lambda dct, tpe, pi, create : CNT(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CNTAnnc, 		CNTAnnc,		lambda dct, tpe, pi, create : CNTAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CNT_LA,		CNT_LA,			lambda dct, tpe, pi, create : CNT_LA(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CNT_OL,		CNT_OL,			lambda dct, tpe, pi, create : CNT_OL(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CSEBase,		CSEBase,		lambda dct, tpe, pi, create : CSEBase(dct, create = create)) 
-addResourceFactoryCallback(T.CSEBaseAnnc,	CSEBaseAnnc,	lambda dct, tpe, pi, create : CSEBaseAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CRS,			CRS,			lambda dct, tpe, pi, create : CRS(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CSR,			CSR,			lambda dct, tpe, pi, create : CSR(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.CSRAnnc,		CSRAnnc,		lambda dct, tpe, pi, create : CSRAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.FCI,			FCI,			lambda dct, tpe, pi, create : FCI(dct, pi = pi, fcntType = tpe, create = create)) 
-addResourceFactoryCallback(T.FCNT,			FCNT,			lambda dct, tpe, pi, create : FCNT(dct, pi = pi, fcntType = tpe, create = create)) 
-addResourceFactoryCallback(T.FCNTAnnc,		FCNTAnnc,		lambda dct, tpe, pi, create : FCNTAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.FCNT_LA,		FCNT_LA,		lambda dct, tpe, pi, create : FCNT_LA(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.FCNT_OL,		FCNT_OL,		lambda dct, tpe, pi, create : FCNT_OL(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.GRP,			GRP,			lambda dct, tpe, pi, create : GRP(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.GRPAnnc,		GRPAnnc,		lambda dct, tpe, pi, create : GRPAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.GRP_FOPT,		GRP_FOPT,		lambda dct, tpe, pi, create : GRP_FOPT(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.NOD,			NOD,			lambda dct, tpe, pi, create : NOD(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.NODAnnc,		NODAnnc,		lambda dct, tpe, pi, create : NODAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.PCH,			PCH,			lambda dct, tpe, pi, create : PCH(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.PCH_PCU,		PCH_PCU,		lambda dct, tpe, pi, create : PCH_PCU(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.REQ,			REQ,			lambda dct, tpe, pi, create : REQ(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.SMD,			SMD,			lambda dct, tpe, pi, create : SMD(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.SMDAnnc,		SMDAnnc,		lambda dct, tpe, pi, create : SMDAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.SUB,			SUB,			lambda dct, tpe, pi, create : SUB(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TS,			TS,				lambda dct, tpe, pi, create : TS(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TSAnnc,		TSAnnc,			lambda dct, tpe, pi, create : TSAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TS_LA,			TS_LA,			lambda dct, tpe, pi, create : TS_LA(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TS_OL,			TS_OL,			lambda dct, tpe, pi, create : TS_OL(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TSB,			TSB,			lambda dct, tpe, pi, create : TSB(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TSI,			TSI,			lambda dct, tpe, pi, create : TSI(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.TSIAnnc,		TSIAnnc,		lambda dct, tpe, pi, create : TSIAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.AE, 			AE,				lambda dct, tpe, pi, create : AE(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.AEAnnc,		AEAnnc,			lambda dct, tpe, pi, create : AEAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CIN, 			CIN,			lambda dct, tpe, pi, create : CIN(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CINAnnc, 		CINAnnc,		lambda dct, tpe, pi, create : CINAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CNT,	 		CNT,			lambda dct, tpe, pi, create : CNT(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CNTAnnc, 		CNTAnnc,		lambda dct, tpe, pi, create : CNTAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CNT_LA,		CNT_LA,			lambda dct, tpe, pi, create : CNT_LA(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CNT_OL,		CNT_OL,			lambda dct, tpe, pi, create : CNT_OL(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CSEBase,		CSEBase,		lambda dct, tpe, pi, create : CSEBase(dct, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CSEBaseAnnc,	CSEBaseAnnc,	lambda dct, tpe, pi, create : CSEBaseAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CRS,			CRS,			lambda dct, tpe, pi, create : CRS(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CSR,			CSR,			lambda dct, tpe, pi, create : CSR(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.CSRAnnc,		CSRAnnc,		lambda dct, tpe, pi, create : CSRAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FCI,			FCI,			lambda dct, tpe, pi, create : FCI(dct, pi = pi, fcntType = tpe, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FCNT,			FCNT,			lambda dct, tpe, pi, create : FCNT(dct, pi = pi, fcntType = tpe, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FCNTAnnc,		FCNTAnnc,		lambda dct, tpe, pi, create : FCNTAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FCNT_LA,		FCNT_LA,		lambda dct, tpe, pi, create : FCNT_LA(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FCNT_OL,		FCNT_OL,		lambda dct, tpe, pi, create : FCNT_OL(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.GRP,			GRP,			lambda dct, tpe, pi, create : GRP(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.GRPAnnc,		GRPAnnc,		lambda dct, tpe, pi, create : GRPAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.GRP_FOPT,		GRP_FOPT,		lambda dct, tpe, pi, create : GRP_FOPT(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.NOD,			NOD,			lambda dct, tpe, pi, create : NOD(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.NODAnnc,		NODAnnc,		lambda dct, tpe, pi, create : NODAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.PCH,			PCH,			lambda dct, tpe, pi, create : PCH(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.PCH_PCU,		PCH_PCU,		lambda dct, tpe, pi, create : PCH_PCU(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.REQ,			REQ,			lambda dct, tpe, pi, create : REQ(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.SMD,			SMD,			lambda dct, tpe, pi, create : SMD(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.SMDAnnc,		SMDAnnc,		lambda dct, tpe, pi, create : SMDAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.SUB,			SUB,			lambda dct, tpe, pi, create : SUB(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TS,			TS,				lambda dct, tpe, pi, create : TS(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TSAnnc,		TSAnnc,			lambda dct, tpe, pi, create : TSAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TS_LA,			TS_LA,			lambda dct, tpe, pi, create : TS_LA(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TS_OL,			TS_OL,			lambda dct, tpe, pi, create : TS_OL(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TSB,			TSB,			lambda dct, tpe, pi, create : TSB(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TSI,			TSI,			lambda dct, tpe, pi, create : TSI(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.TSIAnnc,		TSIAnnc,		lambda dct, tpe, pi, create : TSIAnnc(dct, pi = pi, create = create)) 
 
 # Add for MgmtObj specializations
-addResourceFactoryCallback(T.ANDI,			ANDI,			lambda dct, tpe, pi, create : ANDI(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.ANDIAnnc,		ANDIAnnc,		lambda dct, tpe, pi, create : ANDIAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.ANI,			ANI,			lambda dct, tpe, pi, create : ANI(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.ANIAnnc,		ANIAnnc,		lambda dct, tpe, pi, create : ANIAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.BAT,			BAT,			lambda dct, tpe, pi, create : BAT(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.BATAnnc,		BATAnnc,		lambda dct, tpe, pi, create : BATAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.DVC,			DVC,			lambda dct, tpe, pi, create : DVC(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.DVCAnnc,		DVCAnnc,		lambda dct, tpe, pi, create : DVCAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.DVI,			DVI,			lambda dct, tpe, pi, create : DVI(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.DVIAnnc,		DVIAnnc,		lambda dct, tpe, pi, create : DVIAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.EVL,			EVL,			lambda dct, tpe, pi, create : EVL(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.EVLAnnc,		EVLAnnc,		lambda dct, tpe, pi, create : EVLAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.FWR,			FWR,			lambda dct, tpe, pi, create : FWR(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.FWRAnnc,		FWRAnnc,		lambda dct, tpe, pi, create : FWRAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.MEM,			MEM,			lambda dct, tpe, pi, create : MEM(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.MEMAnnc,		MEMAnnc,		lambda dct, tpe, pi, create : MEMAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.NYCFC,			NYCFC,			lambda dct, tpe, pi, create : NYCFC(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.NYCFCAnnc,		NYCFCAnnc,		lambda dct, tpe, pi, create : NYCFCAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.RBO,			RBO,			lambda dct, tpe, pi, create : RBO(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.RBOAnnc,		RBOAnnc,		lambda dct, tpe, pi, create : RBOAnnc(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.SWR,			SWR,			lambda dct, tpe, pi, create : SWR(dct, pi = pi, create = create)) 
-addResourceFactoryCallback(T.SWRAnnc,		SWRAnnc,		lambda dct, tpe, pi, create : SWRAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.ANDI,			ANDI,			lambda dct, tpe, pi, create : ANDI(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.ANDIAnnc,		ANDIAnnc,		lambda dct, tpe, pi, create : ANDIAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.ANI,			ANI,			lambda dct, tpe, pi, create : ANI(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.ANIAnnc,		ANIAnnc,		lambda dct, tpe, pi, create : ANIAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.BAT,			BAT,			lambda dct, tpe, pi, create : BAT(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.BATAnnc,		BATAnnc,		lambda dct, tpe, pi, create : BATAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.DVC,			DVC,			lambda dct, tpe, pi, create : DVC(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.DVCAnnc,		DVCAnnc,		lambda dct, tpe, pi, create : DVCAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.DVI,			DVI,			lambda dct, tpe, pi, create : DVI(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.DVIAnnc,		DVIAnnc,		lambda dct, tpe, pi, create : DVIAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.EVL,			EVL,			lambda dct, tpe, pi, create : EVL(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.EVLAnnc,		EVLAnnc,		lambda dct, tpe, pi, create : EVLAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FWR,			FWR,			lambda dct, tpe, pi, create : FWR(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.FWRAnnc,		FWRAnnc,		lambda dct, tpe, pi, create : FWRAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.MEM,			MEM,			lambda dct, tpe, pi, create : MEM(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.MEMAnnc,		MEMAnnc,		lambda dct, tpe, pi, create : MEMAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.NYCFC,			NYCFC,			lambda dct, tpe, pi, create : NYCFC(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.NYCFCAnnc,		NYCFCAnnc,		lambda dct, tpe, pi, create : NYCFCAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.RBO,			RBO,			lambda dct, tpe, pi, create : RBO(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.RBOAnnc,		RBOAnnc,		lambda dct, tpe, pi, create : RBOAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.SWR,			SWR,			lambda dct, tpe, pi, create : SWR(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.SWRAnnc,		SWRAnnc,		lambda dct, tpe, pi, create : SWRAnnc(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.WIFIC,			WIFIC,			lambda dct, tpe, pi, create : WIFIC(dct, pi = pi, create = create)) 
+addResourceFactoryCallback(ResourceTypes.WIFICAnnc,		WIFICAnnc,		lambda dct, tpe, pi, create : WIFICAnnc(dct, pi = pi, create = create)) 
 
 
-_specResources = [ T.FCNT, T.FCNTAnnc, T.FCI, T.MGMTOBJ, T.MGMTOBJAnnc ]
+_specResources = [ ResourceTypes.FCNT, ResourceTypes.FCNTAnnc, ResourceTypes.FCI, ResourceTypes.MGMTOBJ, ResourceTypes.MGMTOBJAnnc ]
 
 
-def resourceFromDict(resDict:JSON = {}, pi:str = None, ty:T = None, create:bool = False, isImported:bool = False) -> Result:
+def resourceFromDict(resDict:JSON = {}, pi:str = None, ty:ResourceTypes = None, create:bool = False, isImported:bool = False) -> Result:
 	""" Create a resource from a dictionary structure.
 
 		This function will **not** call the resource's *activate()* method, therefore some attributes
@@ -160,9 +168,9 @@ def resourceFromDict(resDict:JSON = {}, pi:str = None, ty:T = None, create:bool 
 
 		Args:
 			resDict: Dictionary with the resource definition.
-			pi: Resource ID of the parent ID.
-			ty: Resource type.
-			create: Resource will be newly created.
+			pi: The parent's resource ID.
+			ty: The resource type of the resource that shall be created.
+			create: The resource will be newly created.
 			isImported: True when the resource is imported, or created by the `ScriptManager`. In this case some checks may not be performed.
 		Return:
 			`Result` object with the *resource* attribute set to the created resource object.
@@ -171,8 +179,8 @@ def resourceFromDict(resDict:JSON = {}, pi:str = None, ty:T = None, create:bool 
 	resDict, tpe = Utils.pureResource(resDict)	# remove optional "m2m:xxx" level
 
 	# Determine type
-	typ = T(resDict['ty']) if 'ty' in resDict else ty
-	if typ is None and (typ := T.fromTPE(tpe)) is None:
+	typ = ResourceTypes(resDict['ty']) if 'ty' in resDict else ty
+	if typ is None and (typ := ResourceTypes.fromTPE(tpe)) is None:
 		return Result.errorResult(dbg = L.logWarn(f'cannot determine type for creating the resource: {tpe}'))
 
 	if ty is not None:
@@ -182,7 +190,7 @@ def resourceFromDict(resDict:JSON = {}, pi:str = None, ty:T = None, create:bool 
 			return Result.errorResult(dbg = L.logWarn(f'parameter type ({ty}) and resource type specifier ({tpe}) mismatch'))
 	
 	# Check for Parent
-	if pi is None and typ != T.CSEBase and (not (pi := resDict.get('pi')) or len(pi) == 0):
+	if pi is None and typ != ResourceTypes.CSEBase and (not (pi := resDict.get('pi')) or len(pi) == 0):
 		return Result.errorResult(dbg = L.logWarn(f'pi missing in resource: {tpe}'))
 
 	# store the import status in the original resDict
@@ -192,12 +200,12 @@ def resourceFromDict(resDict:JSON = {}, pi:str = None, ty:T = None, create:bool 
 	# Determine a factory and call it
 	factory:FactoryCallableT = None
 
-	if typ == T.MGMTOBJ:										# for <mgmtObj>
+	if typ == ResourceTypes.MGMTOBJ:										# for <mgmtObj>
 		# mgd = resDict['mgd'] if 'mgd' in resDict else None		# Identify mdg in <mgmtObj>
-		factory = T(resDict['mgd']).resourceFactory()
-	elif typ == T.MGMTOBJAnnc:									# for <mgmtObjA>
+		factory = ResourceTypes(resDict['mgd']).resourceFactory()
+	elif typ == ResourceTypes.MGMTOBJAnnc:									# for <mgmtObjA>
 		# mgd = resDict['mgd'] if 'mgd' in resDict else None		# Identify mdg in <mgmtObj>
-		factory = T(resDict['mgd']).announced().resourceFactory()
+		factory = ResourceTypes(resDict['mgd']).announced().resourceFactory()
 	else:
 		factory = typ.resourceFactory()
 	if factory: