+++ acme/resources/WIFICAnnc.py
@@ -1,0 +1,58 @@
+#
+#	WIFICAnnc.py
+#
+#	(c) 2022 by Andreas Kraft
+#	License: BSD 3-Clause License. See the LICENSE file for further details.
+#
+#	WIFIC : Announceable variant
+#
+
+from ..etc.Types import AttributePolicyDict, ResourceTypes, JSON
+from ..resources.MgmtObjAnnc import *
+
+
+class WIFICAnnc(MgmtObjAnnc):
+
+	# Attributes and Attribute policies for this Resource Class
+	# Assigned during startup in the Importer
+	_attributes:AttributePolicyDict = {		
+		# Common and universal attributes for announced resources
+		'rn': None,
+		'ty': None,
+		'ri': None,
+		'pi': None,
+		'ct': None,
+		'lt': None,
+		'et': None,
+		'lbl': None,
+		'acpi':None,
+		'daci': None,
+		'ast': None,
+		'lnk': None,
+
+		# MgmtObj attributes
+		'mgd': None,
+		'obis': None,
+		'obps': None,
+		'dc': None,
+		'mgs': None,
+		'cmlk': None,
+
+		# Resource attributes
+		'ssid': None,
+		'wcrds': None,
+		'maca': None,
+		'chanl': None,
+		'cons': None,
+		'scan': None,
+		'scanr': None,
+		'ud': None,
+		'uds': None,
+		'trdst': None,
+		'rdst': None,
+	}
+
+
+	def __init__(self, dct:JSON = None, pi:str = None, create:bool = False) -> None:
+		super().__init__(dct, pi, mgd = ResourceTypes.WIFICAnnc, create = create)
+